+++ CHANGELOG.md
@@ -9,13 +9,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 <!-- Keep this up to date! After a release, change the tag name to the latest release -->
 [unreleased changes details]: https://github.com/Adobe-Consulting-Services/acs-aem-commons/compare/acs-aem-commons-4.7.2...HEAD
 
+### Added
+- #2518 - Extended the I18N provider / injector mechanism with more options 
+- #2451 - Adding a new dispatcher cache control header filter based on the resource type of the page
+- #2535 - Add option to append new data to arrays using the data importer tool in MCP
+
 ### Fixed
-- 
+- #2529 - Unable to find an implementation for interface acscommons.io.jsonwebtoken.io.Serializer using java.util.ServiceLoader
+- #2535 - Fix issue where when using dry-run functionality in the data importer would still commit the changes
+- #2542 - Fixed issue where VanityUrlAdjuster as in an internal package (and thus could never be implemented)
 
 ## 4.11.2 - 2021-01-05
 
 ### Fixed
 - #2496 - Upgrade shaded Guava dependency to 30.1
+- #2498 - Potential NPE in RunnableOnMaster
+- #2492 - NPE in JcrPackageReplicationStatusEventHandler
 - #2494 - Fixed issue with Versioned ClientLib incompatibility on 6.5.7
 
 ## 4.11.0 - 2020-12-11